Create guess_the_number.java
Generate an integer random number between 1 to 10. Prompt the user to enter a number continuously until the number matches the randomly generated number. For each user input, the program tells the user whether the input is too low or too high, so the user can make the next guess accordingly.